map tool refactor complete
